+++ quantum_debugger/algorithms/__init__.py
@@ -189,6 +189,7 @@ from .adiabatic import adiabatic_evolution
 from .krylov import krylov_spectrum, krylov_ground_energy
 from .loschmidt import loschmidt_echo, rate_function, quench_dynamics
 from .otoc import otoc, scrambling_time
+from .entanglement_growth import entanglement_growth
 from .spectroscopy import unitary_eigenphase, hermitian_eigenvalue
 
 __all__ = [
@@ -383,6 +384,7 @@ __all__ = [
     "quench_dynamics",
     "otoc",
     "scrambling_time",
+    "entanglement_growth",
     # Quantum spectroscopy
     "unitary_eigenphase",
     "hermitian_eigenvalue",